fix, refactor: переписать SQL запросы добавив JOIN, убрать реализацию хранения данных в памяти приложения (InMemory), разделить составные тесты на отдельные части, откорректировать структуру пакетов
