+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.IncorrectPathVariableException;
 import ru.yandex.practicum.filmorate.exception.IncorrectRequestParameterException;
@@ -23,7 +22,7 @@ public class FilmController {
     private final FilmService filmService;
 
     @Autowired
-    public FilmController(@Qualifier("FilmDbService") FilmService filmService) {
+    public FilmController(FilmService filmService) {
         this.filmService = filmService;
     }
