+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.exception.IncorrectPathVariableException;
 import ru.yandex.practicum.filmorate.model.Genre;
-import ru.yandex.practicum.filmorate.service.dao.GenreDbService;
+import ru.yandex.practicum.filmorate.service.GenreDbService;
 
 import java.util.List;
