+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.exception.IncorrectPathVariableException;
 import ru.yandex.practicum.filmorate.model.Mpa;
-import ru.yandex.practicum.filmorate.service.dao.MpaDbService;
+import ru.yandex.practicum.filmorate.service.MpaDbService;
 
 import java.util.List;
