+++ src/main/java/ru/yandex/practicum/filmorate/dao/FilmDbStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.dao;
+package ru.yandex.practicum.filmorate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,11 +8,10 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.exception.AlreadyExistException;
 import ru.yandex.practicum.filmorate.exception.IllegalIdException;
-import ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery;
+import ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
-import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -21,8 +20,8 @@ import java.util.*;
 import static ru.yandex.practicum.filmorate.exception.AlreadyExistException.FILM_ALREADY_EXIST_ADVICE;
 import static ru.yandex.practicum.filmorate.exception.AlreadyExistException.FILM_ALREADY_EXIST_MESSAGE;
 import static ru.yandex.practicum.filmorate.exception.IllegalIdException.*;
-import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery.FILM_INVALID_DATA_BASE_QUERY_ADVICE;
-import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery.INVALID_DATA_BASE_QUERY_MESSAGE;
+import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException.FILM_INVALID_DATA_BASE_QUERY_ADVICE;
+import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException.INVALID_DATA_BASE_QUERY_MESSAGE;
 import static ru.yandex.practicum.filmorate.query.SqlQuery.*;
 
 @Repository
@@ -155,7 +154,7 @@ public class FilmDbStorage implements FilmStorage {
         } catch (EmptyResultDataAccessException exception) {
             log.debug("{}: " + INVALID_DATA_BASE_QUERY_MESSAGE + " Размер ответа на запрос: "
                     + exception.getExpectedSize(), IllegalIdException.class.getSimpleName());
-            throw new InvalidDataBaseQuery(INVALID_DATA_BASE_QUERY_MESSAGE,
+            throw new InvalidDataBaseQueryException(INVALID_DATA_BASE_QUERY_MESSAGE,
                     exception.getExpectedSize(),
                     FILM_INVALID_DATA_BASE_QUERY_ADVICE);
         }
