+++ src/main/java/ru/yandex/practicum/filmorate/dao/GenreDbStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.dao;
+package ru.yandex.practicum.filmorate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,15 +6,15 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.exception.IllegalIdException;
-import ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery;
+import ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException;
 import ru.yandex.practicum.filmorate.model.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
-import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery.GENRE_INVALID_DATA_BASE_QUERY_ADVICE;
-import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQuery.INVALID_DATA_BASE_QUERY_MESSAGE;
+import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException.GENRE_INVALID_DATA_BASE_QUERY_ADVICE;
+import static ru.yandex.practicum.filmorate.exception.InvalidDataBaseQueryException.INVALID_DATA_BASE_QUERY_MESSAGE;
 import static ru.yandex.practicum.filmorate.query.SqlQuery.SQL_QUERY_GET_ALL_GENRES;
 import static ru.yandex.practicum.filmorate.query.SqlQuery.SQL_QUERY_GET_GENRE_BY_ID;
 
@@ -38,7 +38,7 @@ public class GenreDbStorage {
         } catch (EmptyResultDataAccessException exception) {
             log.debug("{}: " + INVALID_DATA_BASE_QUERY_MESSAGE + " Размер ответа на запрос: "
                     + exception.getExpectedSize(), IllegalIdException.class.getSimpleName());
-            throw new InvalidDataBaseQuery(INVALID_DATA_BASE_QUERY_MESSAGE,
+            throw new InvalidDataBaseQueryException(INVALID_DATA_BASE_QUERY_MESSAGE,
                     exception.getExpectedSize(),
                     GENRE_INVALID_DATA_BASE_QUERY_ADVICE);
         }
