+++ src/main/java/ru/yandex/practicum/filmorate/exception/InvalidDataBaseQueryException.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.exception;
 
 import org.springframework.dao.EmptyResultDataAccessException;
 
-public class InvalidDataBaseQuery extends EmptyResultDataAccessException {
+public class InvalidDataBaseQueryException extends EmptyResultDataAccessException {
     public static final String INVALID_DATA_BASE_QUERY_MESSAGE = "Не корректный ответ на запрос к базе данных.";
     public static final String USER_INVALID_DATA_BASE_QUERY_ADVICE = "Проверьте корректность введённых данных. " +
             "База данных присылает пустой ответ на ваш запрос. Возможно вы ввели несуществуюший id пользователя.";
@@ -16,7 +16,7 @@ public class InvalidDataBaseQuery extends EmptyResultDataAccessException {
     // совет пользователю при возникновении исключения:
     private final String adviceToUser;
 
-    public InvalidDataBaseQuery(String msg, int expectedSize, String adviceToUser) {
+    public InvalidDataBaseQueryException(String msg, int expectedSize, String adviceToUser) {
         super(msg, expectedSize);
         this.adviceToUser = adviceToUser;
     }
