+++ src/main/java/ru/yandex/practicum/filmorate/service/UserDbService.java
@@ -1,19 +1,14 @@
-package ru.yandex.practicum.filmorate.service.dao;
+package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.dao.UserDbStorage;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.service.UserService;
-import ru.yandex.practicum.filmorate.storage.dao.UserDbStorage;
 
 import java.util.List;
 
 @Service
-@Qualifier("UserDbService")
 public class UserDbService implements UserService {
-    private static final String REMOVE_FROM_FRIEND_MESSAGE = "Пользователи успешно удалены из друзей. Их id: ";
-
     private final UserDbStorage userStorage;
 
     @Autowired
